split asset_widgets into component files
