+++ lib/core/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:g4c/domain/use_cases/personality_quiz_runner.dart';
 // ignore: unused_import
-import '../presentation/components/asset_widgets.dart';
 import 'package:g4c/presentation/views/login.dart';
 import 'package:firebase_core/firebase_core.dart';
 import '../data/data_sources/firebase_options.dart';
