+++ lib/domain/repositories/enumerations.dart
@@ -1,0 +1,1 @@
+enum Preference { disagree, avoid, neutral, tolerate, agree }
