+++ lib/presentation/views/profile_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:g4c/presentation/components/asset_widgets.dart';
+import 'package:g4c/presentation/components/card_widget.dart';
 import 'package:g4c/presentation/components/g4c_drawer.dart';
+import 'package:g4c/presentation/components/top_card.dart';
 
 class ProfilePage extends StatelessWidget {
   final double fontsize = 15.0;
