+++ pubspec.yaml
@@ -65,7 +65,7 @@ flutter:
   assets:
     - asset_lib/images/
     - asset_lib/images/roleimages/
-    - asset_lib/json/
+    - lib/data/data_sources/json/
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
